reword! Silence misguided GCC warning about alignment issues
new Subject:
libarchive: avoid type punning corner case

When compiling xz_dec_stream.c with GCC (at least with versions 5.4.0
and 7.1.0), it complains thusly:

        In function 'dec_stream_footer':
        error: dereferencing type-punned pointer will break
              strict-aliasing rules [-Werror=strict-aliasing]
           if (xz_crc32(s->temp.buf + 4, 6, 0) != get_le32(s->temp.buf))
               ^~

Despite the arrow pointing to the xz_crc32() call, the actual problem is
the get_le32() macro expansion.

The `buf` field is an array of uint8_t. In this case, the macro expansion
amounts to a cast of the array to a pointer to an unrelated type. It looks
like GCC does not decay the array to a pointer before it applies the cast.

We could just write the sub-expression as get_le32(s->temp.buf + 0), which
forces the array to decay to a pointer. But future readers would have to
be enlightened about the purpose the + 0 with a comment.

Let's do it differently: Turn the get_le32() macro into an inline
function taking a pointer parameter. Then the array-to-pointer decay
happens as part of the argment passing and without funny gymnastics.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Johannes Sixt <j6t@kdbg.org>