Add a GitHub workflow
This not only ensures that BusyBox-w32 can be built directly on Windows
(in an MSYS2 environment using mingw-w64-gcc), it also ensures that the
testsuite runs and passes.

Note: we need the `build-installers` flavor of Git for Windows' SDK
because the smaller variants lack tools like `bzip2` and `od` that are
required by BusyBox' own build (apparently they are required when
splitting the config, for some reason).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>