ci: also run the test suite
Now that we added fixes or workarounds for all failures when running
with MSYS2, let's run it in CI builds, to keep it in this shape.

Note that we use BusyBox itself to run the test suite. The original plan
was to use MSYS2's Bash and coreutils, so that the output of BusyBox
could be validated against the MSYS2 tools' output, but it turned out
that in particular the MSYS2 path translation (where it pretends to have
a Unix root directory, and tries to convert environment variables and
command-line arguments to Windows paths as necessary, sometimes
mistaking e.g. regular expressions for being Unix paths) was simply too
hairy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>