tests(od): skip a test case that won't ever work as expected on Windows
`od`'s test assumes that scientific notation will be output with two
digits for the exponent. However, on Windows that is not so. There
_used_ to be a function called `_set_output_format()` that allowed
controlling the format, but that function is gone.

For more on this, see
https://stackoverflow.com/questions/31331723/how-to-control-the-number-of-exponent-digits-after-e-in-c-printf-e

In light of this, make this particular test case contingent on
`PLATFORM_POSIX`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>