tests(printf): avoid using $HOME
The home directory on Windows contains backslashes. In the `testing`
framework, these characters can be mistaken for something they are not.
For example, when running the tests in GitHub workflows, the home
directory might be something like `C:\Users\runneradmin`, and that `\r`
would be mistaken for a newline.

It is not at all necessary to use `$HOME` here, anyway. Any old string
will do.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>