Do not treat symlink failures as mere warnings
In 7587b56c1 (win32: implement symlink(2), 2021-02-12), we started
issuing only a warning when a symbolic link could not be created,
instead of dying.

However, this breaks the expectation of the `tar` test case "Symlink
attack: create symlink and then write through it" which explictly
verifies that the exit code of `tar xvf ...` is 1 when it failed to
create a symbolic link.

So let's exit with an error in that instance.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>