mingw: allow using a MINGW bzip2.exe
The MINGW variant does not understand `/dev/null`, and therefore `bzip2`
complains thusly:

	bzip2.exe: I won't write compressed data to a terminal.
	bzip2.exe: For help, type: `bzip2.exe --help'.

and then the script incorrectly assumes that `bzip2` is not even installed.

Let's instead use the common pattern of showing the usage and testing
for the exit code 127 (which means that the command was not found).

While at it, give `od` the same treatment.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>