win32/timegm.c: fix incorrect and convoluted code
In 9e16eecc7 (win32: changes to allow timezones in dates, 2021-09-17), a
horrible piece of code was introduced where the `is_leap` pointer was
set to `&(int){0}` if it previously was `NULL`.

The idea most likely being: we can avoid introducing a local,
potentially unused variable by defining an unnamed struct with an
implicit `int` member and then using the pointer to that struct.

This is not only almost unreadable and poor practice, it is also
something that GCC considers a bug because the `{0}` construct
implicitly defines an unnamed struct whose scope is the conditional
block of the `if (!is_leap)`, i.e. it is used and modified when already
out of scope.

Consequently, GCC warns loudly about it:

	win32/timegm.c: In function '__year_to_secs':
	win32/timegm.c:74:45: error: dangling pointer 'is_leap' to an unnamed temporary may be used [-Werror=dangling-pointer=]
	   74 |         leaps += 97*cycles + 24*centuries - *is_leap;
	      |                                             ^~~~~~~~
	win32/timegm.c:45:39: note: unnamed temporary defined here
	   45 |         if (!is_leap) is_leap = &(int){0};
	      |                                       ^

Let's fix this in the traditional, correct way, by using a temporary
local variable that has the correct scope. And is eminently more
readable.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>