mingw: fix cp-RHL-does_not_preserve-links (again)
In 367f583cf (win32: let `resolve_symlinks()` _really_ resolve symbolic
links, 2021-02-24), I fixed a problem where symbolic links were
dereferenced incorrectly (if `a/link` points to `file`, it would be
dereferenced to `./file` instead of the correct `a/file`).

But 31467ddfc (win32: changes to stat(2) implementation, 2022-05-23)
broke that again (which would have been caught if there were any CI
builds worth mentioning).

Let's just revert the latter. It is better to have code that does what
it is supposed to do instead of "better" code that does not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>