ash: only "eat" Carriage Returns preceding Line Feeds
In 6a798baff (Eat trailing CRs as well as LFs when expanding backquotes,
2012-02-16), a change was introduced where all Carriage Returns were
eaten when expanding backquotes, regardless where in the input they
happen.

This is incorrect, though, as `x="$(printf '\015')"` should not assign
the empty string to the variable, but a single Carriage Return.

This is needed to let Git's t5100.50 "mailinfo warn CR in base64 encoded
email" succeed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>