normalize_ntpath(): do _not_ convert slashes
We will need a pure version in the next commits that normalizes the NT
paths without converting the slashes.

While at it, move the function a little up so that we can use it in the
upcoming `wreadlink()` function whose location is dictated by its users.

This commit is best viewed with `--color-moved`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>